@@ -2,46 +2,47 @@
     "private": true,
     "scripts": {
         "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --config=node_modules/laravel-mix/setup/webpack.config.js",
         "watch": "npm run development -- --watch",
         "watch-poll": "npm run watch -- --watch-poll",
         "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
         "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --config=node_modules/laravel-mix/setup/webpack.config.js"
     },
     "devDependencies": {
         "@fortawesome/fontawesome-free": "^5.11.2",
-        "@types/crypto-js": "^3.1.43",
-        "@types/socket.io-client": "^1.4.32",
-        "axios": "^0.18.1",
-        "bootstrap": "^4.3.1",
-        "bootstrap-vue": "^2.0.4",
-        "cross-env": "^5.2.1",
-        "crypto-js": "^3.1.9-1",
-        "dotenv": "^8.2.0",
-        "enhanced-resolve": "^4.1.1",
-        "jquery": "^3.4.1",
-        "laravel-mix": "^4.1.4",
-        "localforage": "^1.7.3",
-        "lodash": "^4.17.15",
-        "ndef": "^0.2.0",
-        "node-fetch": "^2.6.0",
-        "popper.js": "^1.16.0",
-        "resolve-url-loader": "^2.3.1",
-        "sass": "^1.23.0",
-        "sass-loader": "^7.3.1",
-        "socket.io-client": "^2.3.0",
-        "swagger-ui": "^3.24.0",
-        "ts-loader": "^6.2.0",
-        "typescript": "^3.6.4",
-        "uuid": "^3.3.3",
-        "vue": "^2.6.10",
-        "vue-router": "^3.1.3",
-        "vue-template-compiler": "^2.6.10"
-    },
-    "dependencies": {
+        "@types/crypto-js": "^4.2.2",
+        "@types/uuid": "^10.0.0",
+        "@vue/compat": "^3.5.13",
+        "@vue/compiler-sfc": "^3.5.13",
         "airbrake-js": "^2.0.0-beta.3",
-        "moment": "^2.24.0",
-        "vue-cookies": "^1.6.1"
+        "axios": "^1.13.5",
+        "bootstrap-vue": "^2.23.1",
+        "buffer": "^6.0.3",
+        "cross-env": "^7.0.3",
+        "crypto-js": "^4.2.0",
+        "dotenv": "^16.4.7",
+        "enhanced-resolve": "^5.18.0",
+        "jquery": "^3.7.1",
+        "laravel-mix": "^6.0.49",
+        "localforage": "^1.10.0",
+        "lodash": "^4.17.21",
+        "moment": "^2.30.1",
+        "ndef": "^0.2.0",
+        "node-fetch": "^3.3.2",
+        "qrcode.vue": "^3.8.0",
+        "resolve-url-loader": "^5.0.0",
+        "sass": "^1.83.1",
+        "sass-loader": "^16.0.4",
+        "socket.io-client": "^4.8.1",
+        "swagger-ui": "^5.18.2",
+        "ts-loader": "^9.5.1",
+        "typescript": "^5.7.2",
+        "uuid": "^11.0.3",
+        "vue": "^3.5.13",
+        "vue-cookies": "^1.8.5",
+        "vue-loader": "^16.8.3",
+        "vue-qrcode-reader": "^5.7.3",
+        "vue-router": "^4.5.0"
     }
 }
