@@ -22,9 +22,11 @@
 
 namespace App\Exceptions;
 
+use CatLab\Charon\Laravel\Exceptions\CharonErrorHandler;
 use Exception;
-use Http\Client\Exception\HttpException;
+use HttpException;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
+use Throwable;
 
 class Handler extends ExceptionHandler
 {
@@ -34,7 +36,7 @@ class Handler extends ExceptionHandler
      * @var array
      */
     protected $dontReport = [
-        HttpException::class
+        \Http\Message\Exception::class
     ];
 
     /**
@@ -53,7 +55,7 @@ class Handler extends ExceptionHandler
      * @param  \Exception  $exception
      * @return void
      */
-    public function report(Exception $exception)
+    public function report(Throwable $exception)
     {
         parent::report($exception);
     }
@@ -65,8 +67,15 @@ class Handler extends ExceptionHandler
      * @param  \Exception  $exception
      * @return \Illuminate\Http\Response
      */
-    public function render($request, Exception $exception)
+    public function render($request, Throwable $exception)
     {
+        // Try to use Charons  handler
+        $charonHandler = new CharonErrorHandler();
+        $response = $charonHandler->handleException($request, $exception);
+        if ($response) {
+            return $response;
+        }
+
         return parent::render($request, $exception);
     }
 }
